+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Version 0.4.1.0
 [+] added creationTimestamp for IEvent
 
+[+] added timezone offset obtaining method 
+
 ### Version 0.4.0.0
 [*] dataLen -> dataSize